refactor: moved injection package to the core module
